+++ src/modules/profile/EditProfileView.js
@@ -3,30 +3,15 @@ import View from 'react-flux-state';
 import { toast } from 'react-toastify';
 import * as R from 'ramda';
 import { Link } from 'react-router-dom';
-import { MDBCol, MDBContainer, MDBRow, MDBIcon, MDBAnimation } from 'mdbreact';
+import { MDBContainer, MDBRow, MDBIcon, MDBAnimation } from 'mdbreact';
 import { landingStore, USER_EVENT } from '../landing/landing-store';
 import { inventoryStore, INVENTORY_EVENT } from '../inventory/inventory-store';
-import {
-  profileStore,
-  ACCOUNT_ERROR_EVENT,
-  NEW_ACCOUNT_EVENT,
-  DELETE_ACCOUNT_EVENT,
-  UPDATE_ACCOUNT_EVENT,
-} from './profile-store';
-import {
-  addAccountAction,
-  updateAccountAction,
-  fetchProfileAction,
-  updateProfileAction,
-} from './profile-actions';
+import { fetchProfileAction, updateProfileAction } from './profile-actions';
 import { fetchUserProducts } from '../inventory/inventory-actions';
 import { userModel } from './Profile-models';
-import { ModalConfirm } from '../../components/ModalConfirm';
 import { EditBasicInformation } from './components/EditBasicInformation';
-import { EditBankInformation } from './components/EditBankInformation';
 import { Loader } from '../../components/Loader';
 import SidebarComponent from '../../components/SidebarComponent';
-import SliderCardsMap from '../../components/SliderCardsMap';
 
 class EditProfileView extends View {
   constructor(props) {
@@ -53,11 +38,6 @@ class EditProfileView extends View {
       });
     });
 
-    this.subscribe(profileStore, ACCOUNT_ERROR_EVENT, (e) => {
-      toast.error(e.message);
-      this.setState({ loadingBankAccounts: false });
-    });
-
     this.subscribe(landingStore, USER_EVENT, (user) => {
       this.setState(
         {
@@ -71,86 +51,9 @@ class EditProfileView extends View {
         },
       );
     });
-
-    this.subscribe(profileStore, ACCOUNT_ERROR_EVENT, (e) => {
-      toast.error(e.message);
-      this.setState({ loadingBankAccounts: false });
-    });
-
-    this.subscribe(profileStore, NEW_ACCOUNT_EVENT, (bankAccount) => {
-      const { user } = this.state;
-      user.bankAccounts.push(bankAccount);
-      this.setState(
-        {
-          loadingBankAccounts: false,
-          user,
-        },
-        () => {
-          toast.info('New Account Successfully Created');
-        },
-      );
-    });
-
-    this.subscribe(profileStore, DELETE_ACCOUNT_EVENT, (index) => {
-      const { user } = this.state;
-      user.bankAccounts.splice(index, 1);
-      this.setState(
-        {
-          loadingBankAccounts: false,
-          user,
-        },
-        () => {
-          toast.info('Account Deleted');
-        },
-      );
-    });
-
-    this.subscribe(profileStore, UPDATE_ACCOUNT_EVENT, (bankAccount) => {
-      let { user } = this.state;
-      let index = bankAccount[1];
-      user.bankAccounts[index] = bankAccount[0];
-      this.setState(
-        {
-          loadingBankAccounts: false,
-          user,
-        },
-        () => {
-          toast.info('Account Updated');
-        },
-      );
-    });
-
     fetchUserProducts();
     fetchProfileAction();
   }
-
-  newAccount = (account) => {
-    this.setState({ loadingBankAccounts: true }, () => {
-      addAccountAction({ ...account });
-    });
-  };
-
-  editAccount = (account, index) => {
-    this.setState({ loadingBankAccounts: true }, () => {
-      updateAccountAction({ ...account }, index);
-    });
-  };
-
-  onDeleteBankAccount = (index) => {
-    this.setState((prevState) => ({
-      deleteBankAccountModalIsOpen: !prevState.deleteBankAccountModalIsOpen,
-      bankAccountIndex: index,
-    }));
-  };
-
-  deleteBankAccount = () => {
-    this.setState({ loadingBankAccounts: true }, () => {
-      const user = R.clone(this.state.user);
-      const { bankAccounts } = user;
-      bankAccounts.splice(this.state.bankAccountIndex, 1);
-      updateProfileAction(user);
-    });
-  };
 
   onUpdateUser = (updateUser) => {
     const user = R.clone(this.state.user);
@@ -173,14 +76,8 @@ class EditProfileView extends View {
   };
 
   render() {
-    const { name, description, picture, bankAccounts } = this.state.user;
-    const {
-      loadingBankAccounts,
-      loadingUser,
-      loadingInventory,
-      deleteBankAccountModalIsOpen,
-      inventory,
-    } = this.state;
+    const { name, description, picture } = this.state.user;
+    const { loadingUser } = this.state;
     return (
       <SidebarComponent>
         <div className="d-flex justify-content-between nav-admin body">
@@ -212,47 +109,6 @@ class EditProfileView extends View {
             ) : (
               <div></div>
             )}
-            <hr />
-            <MDBRow>
-              <MDBCol md="3" />
-              <MDBCol md="9">
-                {loadingBankAccounts ? (
-                  <div className="text-center">
-                    <Loader />
-                  </div>
-                ) : (
-                  <div>
-                    <ModalConfirm
-                      open={deleteBankAccountModalIsOpen}
-                      onClose={this.toggleModal}
-                      text={'Are you sure you want to Delete the Bank Account?'}
-                      onOk={this.deleteBankAccount}
-                    />
-                    <EditBankInformation
-                      editAccount={this.editAccount}
-                      newAccount={this.newAccount}
-                      bankAccounts={bankAccounts}
-                      onDelete={this.onDeleteBankAccount}
-                    />
-                  </div>
-                )}
-              </MDBCol>
-            </MDBRow>
-            <MDBRow>
-              <MDBCol md="3" />
-              <MDBCol md="9">
-                <h5 className="font-weight-bold text-black-50">
-                  Recently Publications
-                </h5>
-                {loadingInventory ? (
-                  <div className="text-center">
-                    <Loader />
-                  </div>
-                ) : (
-                  <SliderCardsMap inventory={inventory} />
-                )}
-              </MDBCol>
-            </MDBRow>
           </MDBContainer>
         </MDBAnimation>
       </SidebarComponent>
